moved pype info widget to separate file
