+++ pype/tools/tray/pype_tray.py
@@ -3,60 +3,12 @@ import sys
 
 import platform
 from avalon import style
-from Qt import QtCore, QtGui, QtWidgets, QtSvg
+from Qt import QtCore, QtGui, QtWidgets
 from pype.api import Logger, resources
 from pype.modules import TrayModulesManager, ITrayService
 from pype.settings.lib import get_system_settings
 import pype.version
-
-
-class PypeInfoWidget(QtWidgets.QWidget):
-    not_allowed = "N/A"
-
-    def __init__(self, parent=None):
-        super(PypeInfoWidget, self).__init__(parent)
-
-        self.setStyleSheet(style.load_stylesheet())
-
-        icon = QtGui.QIcon(resources.pype_icon_filepath())
-        self.setWindowIcon(icon)
-        self.setWindowTitle("Pype info")
-
-        main_layout = QtWidgets.QFormLayout(self)
-        main_layout.setLabelAlignment(QtCore.Qt.AlignRight)
-        if getattr(sys, "frozen", False):
-            version_end = "build"
-        else:
-            version_end = "code"
-        version_value = "{} ({})".format(
-            pype.version.__version__, version_end
-        )
-
-        lable_value = [
-            # Pype version
-            ("Pype version:", version_value),
-            ("Pype location:", os.environ.get("PYPE_ROOT")),
-
-            # Mongo URL
-            ("Pype Mongo URL:", os.environ.get("PYPE_MONGO"))
-        ]
-
-        for label, value in lable_value:
-            main_layout.addRow(
-                label,
-                QtWidgets.QLabel(value or self.not_allowed)
-            )
-
-    def showEvent(self, event):
-        result = super(PypeInfoWidget, self).showEvent(event)
-        screen_center = (
-            QtWidgets.QApplication.desktop().availableGeometry(self).center()
-        )
-        self.move(
-            screen_center.x() - (self.width() / 2),
-            screen_center.y() - (self.height() / 2)
-        )
-        return result
+from .pype_info_widget import PypeInfoWidget
 
 
 class TrayManager:
